[FIX] Fix OS_LoggerFormat's vtable misplacement
OS_LoggerFormat_Handle_t's vtable was on the bottom of the structure
wherein the parent's OS_LoggerAbstractFormat_Handle it is the first
member.

This was not an issue, as all over the code the child class is passed
and cast, but it must be fixed before it becomes a true issue.

Fixes:
SEOS-1388