[NEW] Add error code to string_converter
It was hard to choose the proper implementation, and it was very
tempting to use some powerful macro magic but in the end I've chosen
simple ternary operators cascade. It works, and compiler/linker should
be able to optimize it.

Some complicated macro magic will definitely cause huge confusions at
the first glance, and would also probably require large lookup table
(at least 1318 bytes).

Furthermore non-static function with a switch case would require
creating a new library for it, as we are not allowed to place source
files in the core api.

Other options that were considered:
<https://stackoverflow.com/questions/9907160/how-to-convert-enum-names-to-string-in-c>

Fixes:
SEOS-2433